requirements.txt
discord.py
gtts
pyogg
aiohttp